agents/data_agent/requirements.txt
flask
flask-cors
psycopg2-binary
requests
pandas==2.0.3
numpy==1.24.4
